LayoutTests/fast/events/popup-blocking-timers4-expected.txt

Click Here
Test calling window.open() in a nested call to setTimeout(). A popup should not be allowed.
PASS newWindow is undefined.
